Extract diagnostics peak table builder
