+++ apps/server/tests/adapters/pdf/test_report_analysis_findings_integration.py
@@ -17,7 +17,9 @@ from vibesensor.use_cases.diagnostics import findings as findings_builder_module
 from vibesensor.use_cases.diagnostics.findings import _build_findings as _findings_build_findings
 from vibesensor.use_cases.diagnostics.findings import _speed_breakdown
 from vibesensor.use_cases.diagnostics.location_analysis import LocationAnalysisResult
-from vibesensor.use_cases.diagnostics.plots import top_peaks_table_rows as _top_peaks_table_rows
+from vibesensor.use_cases.diagnostics.peak_table import (
+    top_peaks_table_rows as _top_peaks_table_rows,
+)
 
 
 def test_speed_breakdown_basic() -> None:
