+++ apps/server/tests/adapters/pdf/test_report_floor_estimation.py
@@ -6,7 +6,9 @@ from vibesensor.use_cases.diagnostics.helpers import (
     _estimate_strength_floor_amp_g,
     _run_noise_baseline_g,
 )
-from vibesensor.use_cases.diagnostics.plots import top_peaks_table_rows as _top_peaks_table_rows
+from vibesensor.use_cases.diagnostics.peak_table import (
+    top_peaks_table_rows as _top_peaks_table_rows,
+)
 from vibesensor.vibration_strength import percentile
