+++ apps/server/tests/adapters/pdf/test_report_persistence_classification.py
@@ -5,7 +5,7 @@ from _report_persistence_helpers import build_findings, findings_at_freq, unifor
 from test_support.report_helpers import analysis_sample_with_peaks as sample
 
 from vibesensor.use_cases.diagnostics.findings import _classify_peak_type
-from vibesensor.use_cases.diagnostics.plots import (
+from vibesensor.use_cases.diagnostics.peak_table import (
     top_peaks_table_rows as _top_peaks_table_rows,
 )
 from vibesensor.use_cases.diagnostics.spectrogram import (
