+++ apps/server/tests/adapters/pdf/test_report_persistence_summary.py
@@ -9,7 +9,7 @@ from test_support.report_helpers import (
 )
 
 from vibesensor.use_cases.diagnostics import build_findings_for_samples
-from vibesensor.use_cases.diagnostics.plots import (
+from vibesensor.use_cases.diagnostics.peak_table import (
     top_peaks_table_rows as _top_peaks_table_rows,
 )
 from vibesensor.use_cases.diagnostics.spectrogram import (
